+++ src/modules/monapi/collect/collector/mysql.go
@@ -1,25 +1,26 @@
 package collector
 
 import (
-	"encoding/json"
 	"fmt"
 
-	"github.com/didi/nightingale/src/models"
 	"github.com/didi/nightingale/src/modules/monapi/collect"
 	"github.com/didi/nightingale/src/modules/monapi/collect/collector/mysql"
 )
 
 func init() {
-	collect.CollectorRegister(&MysqlCollector{})
+	collect.CollectorRegister(NewMysqlCollector())
 }
 
-type MysqlCollector struct{}
+type MysqlCollector struct {
+	*collect.BaseCollector
+}
 
-func (p MysqlCollector) Name() string               { return "mysql" }
-func (p MysqlCollector) Category() collect.Category { return collect.RemoteCategory }
-
-func (p MysqlCollector) Template() (interface{}, error) {
-	return collect.Template(&MysqlRule{})
+func NewMysqlCollector() *MysqlCollector {
+	return &MysqlCollector{BaseCollector: collect.NewBaseCollector(
+		"mysql",
+		collect.RemoteCategory,
+		func() interface{} { return &MysqlRule{} },
+	)}
 }
 
 type MysqlRule struct {
@@ -46,6 +47,13 @@ type MysqlRule struct {
 	MetricVersion                       int      `label:"-" json:"-"`
 }
 
+func (p *MysqlRule) Validate() error {
+	if len(p.Servers) == 0 || p.Servers[0] == "" {
+		return fmt.Errorf("mysql.rule.servers must be set")
+	}
+	return nil
+}
+
 func (p MysqlRule) Mysql() *mysql.Mysql {
 	return &mysql.Mysql{
 		Servers:                             p.Servers,
@@ -70,133 +78,4 @@ func (p MysqlRule) Mysql() *mysql.Mysql {
 		IntervalSlow:                        p.IntervalSlow,
 		MetricVersion:                       2,
 	}
-}
-
-func (p *MysqlRule) Validate() error {
-	if len(p.Servers) == 0 || p.Servers[0] == "" {
-		return fmt.Errorf("mysql.rule.servers must be set")
-	}
-	return nil
-}
-
-func (p MysqlCollector) Get(id int64) (interface{}, error) {
-	collect := &models.CollectRule{}
-	has, err := models.DB["mon"].Where("id = ?", id).Get(collect)
-	if !has {
-		return nil, err
-	}
-	return collect, err
-}
-
-func (p MysqlCollector) Gets(nids []int64) (ret []interface{}, err error) {
-	collects := []models.CollectRule{}
-	err = models.DB["mon"].In("nid", nids).Find(&collects)
-	for _, c := range collects {
-		ret = append(ret, c)
-	}
-	return ret, err
-}
-
-func (p MysqlCollector) GetByNameAndNid(name string, nid int64) (interface{}, error) {
-	collect := &models.CollectRule{}
-	has, err := models.DB["mon"].Where("collect_type = ? and name = ? and nid = ?", p.Name(), name, nid).Get(collect)
-	if !has {
-		return nil, err
-	}
-	return collect, err
-}
-
-func (p MysqlCollector) Create(data []byte, username string) error {
-	collect := &models.CollectRule{CollectType: p.Name()}
-	rule := &MysqlRule{}
-
-	if err := json.Unmarshal(data, collect); err != nil {
-		return fmt.Errorf("unmarshal body %s err:%v", string(data), err)
-	}
-
-	if err := collect.Validate(rule); err != nil {
-		return err
-	}
-
-	can, err := models.UsernameCandoNodeOp(username, "mon_collect_create", collect.Nid)
-	if err != nil {
-		return err
-	}
-	if !can {
-		return fmt.Errorf("permission deny")
-	}
-
-	collect.Creator = username
-	collect.LastUpdator = username
-
-	old, err := p.GetByNameAndNid(collect.Name, collect.Nid)
-	if err != nil {
-		return err
-	}
-	if old != nil {
-		return fmt.Errorf("同节点下策略名称 %s 已存在", collect.Name)
-	}
-	return models.CreateCollect(p.Name(), username, collect)
-}
-
-func (p MysqlCollector) Update(data []byte, username string) error {
-	collect := &models.CollectRule{}
-	rule := &MysqlRule{}
-
-	if err := json.Unmarshal(data, collect); err != nil {
-		return fmt.Errorf("unmarshal body %s err:%v", string(data), err)
-	}
-
-	if err := collect.Validate(rule); err != nil {
-		return err
-	}
-
-	can, err := models.UsernameCandoNodeOp(username, "mon_collect_modify", collect.Nid)
-	if err != nil {
-		return err
-	}
-	if !can {
-		return fmt.Errorf("permission deny")
-	}
-
-	//校验采集是否存在
-	obj, err := p.Get(collect.Id) //id找不到的情况
-	if err != nil {
-		return fmt.Errorf("采集不存在 type:%s id:%d", p.Name(), collect.Id)
-	}
-
-	tmpId := obj.(*models.CollectRule).Id
-	if tmpId == 0 {
-		return fmt.Errorf("采集不存在 type:%s id:%d", p.Name(), collect.Id)
-	}
-
-	collect.Creator = username
-	collect.LastUpdator = username
-
-	old, err := p.GetByNameAndNid(collect.Name, collect.Nid)
-	if err != nil {
-		return err
-	}
-	if old != nil && tmpId != old.(*models.CollectRule).Id {
-		return fmt.Errorf("同节点下策略名称 %s 已存在", collect.Name)
-	}
-
-	return collect.Update()
-}
-
-func (p MysqlCollector) Delete(id int64, username string) error {
-	tmp, err := p.Get(id) //id找不到的情况
-	if err != nil {
-		return fmt.Errorf("采集不存在 type:%s id:%d", p.Name(), id)
-	}
-	nid := tmp.(*models.CollectRule).Nid
-	can, err := models.UsernameCandoNodeOp(username, "mon_collect_delete", int64(nid))
-	if err != nil {
-		return err
-	}
-	if !can {
-		return fmt.Errorf("permission deny")
-	}
-
-	return models.DeleteCollectById(p.Name(), username, id)
 }
